requirements.txt
bitarray==0.8.1
Pillow==4.2.1
